isolate LearnBase candidates
